chore: moved jwt auth filter from security to config/filter
